Enable all layer for WebGLBackgrond.
Before https://github.com/mrdoob/three.js/pull/22426,
when not using array camera, the background doesn't participate
in layer calculation.

We feel the old behaviour is better because we don't need to
always add layer `0` to the camera when doing multipass rendering.

This commit enables all layer for the background.

Note that the behaviour is still slightly different from r115:

When the camera's layer **mask** is set to 0, r115 will render
background and now won't.